+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Changelog dev site doc page
+
+1.2.0 - (January 2, 2019)
+-----------------
+### Fixed
+* Fixed paging logic when rows are being removed from the DataGrid
 
 1.1.0 - (November 21, 2018)
 -----------------